feat: update search text field
