+++ AppCore/Sources/Screens/Home/HomeViewController.swift
@@ -71,7 +71,7 @@ public final class HomeViewController: UIViewController, TabBarPresentable {
 
     // swiftlint:disable:next function_body_length
     private func setupBinding() {
-        contentView.onChangeSearchQuery = { [presenter] query in
+        contentView.searchTextField.onChangeText = { [presenter] query in
             presenter.searchQueryChanged(query: query)
         }
